@@ -1,15 +1,16 @@
 {
-  "name": "react-redux-saga-boilerplate",
+  "name": "react-redux",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-redux": "^5.0.7",
+    "@material-ui/core": "^4.4.1",
+    "axios": "^0.18.1",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
-    "redux": "^4.0.0",
-    "redux-saga": "^0.16.0"
+    "redux": "^4.0.4",
+    "redux-saga": "^0.16.2"
   },
   "devDependencies": {
     "react-scripts": "1.1.4"
