@@ -1,43 +1,54 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
+import React, {useState } from 'react';
+import { Container, Typography, Card, Grid, TextField, Button } from '@material-ui/core';
+import { typography } from '@material-ui/system';
+import { mergeClasses } from '@material-ui/styles';
 
-import { getDemoRequest } from '../../redux/actions/demoActions';
+import styles from './style';
 
-import User from '../../components/User';
+export default () => {
 
-class Home extends Component {
-	componentWillMount() {
-		this.props.getDemoRequest('hey');
+	const [searchText, setSearchText] = useState('');
+
+	const classes = styles();
+
+	const handleSearchTextChange = event =>{
+		setSearchText(event.target.value);
 	}
-	render() {
-		const { users } = this.props;
 
-		let items = [];
-		if (typeof users !== 'undefined') {
-			items = users.map((value, index) => {
-				return <User key={index} {...value} />;
-			});
-		}
-		return <div>{items}</div>;
+	const handleCleanTextClick = event =>{
+		console.log(10);
 	}
+
+	const handleSearchTextClick = event =>{
+		console.log(10);
+	}
+
+	console.log(searchText);
+
+	return(
+		<Container className={classes.container}>
+			<Card className={classes.cardContainer}>
+				<Grid container className={classes.titleGraidContainer}>
+					<Grid>
+					<Typography className={classes.title}> Bienvenido!</Typography>
+					</Grid>
+					<Grid>
+					<label>
+						Icono
+					</label>
+				</Grid>
+				</Grid>
+				<TextField 
+				value={searchText}
+				placeholder="Buscar ..."
+				className={classes.textFieldSearch}
+				onChange={handleSearchTextChange}/>
+				<Grid className={classes.buttonsContainer}>
+					<Button variant = "contained" onClick={handleCleanTextClick}> Limpiar</Button>
+					<Button variant = "contained" color="primary" size="large" className={classes.searchButtonContainer} onClick={handleSearchTextClick}> Buscar</Button>
+				</Grid>
+
+			</Card>
+		</Container>
+	)
 }
-
-const mapDispatchToProps = (dispatch, props) => {
-	return {
-		getDemoRequest: payload => {
-			dispatch(getDemoRequest(payload));
-		}
-	};
-};
-const mapStateToProps = state => {
-	return {
-		users: state.demoReducer[0]
-	};
-};
-
-Home.propTypes = {
-	dispatch: PropTypes.func
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Home);
